Java Gold勉強進捗 品質改善 ・ファイル構造見直し→ "mvn test"実行可能状態
