+++ src/main/java/chapter4/IntermediateOperations.java
@@ -1,4 +1,4 @@
-package org.chapter4;
+package chapter4;
 
 import java.util.List;
 import java.util.stream.Stream;
