+++ src/main/java/common/ErrorText.java
@@ -1,4 +1,4 @@
-package org.common;
+package common;
 
 /**
  * 想定エラーのケース
