+++ src/main/java/common/annotation/TestFlag.java
@@ -1,0 +1,14 @@
+package common.annotation;
+
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+@Retention(RetentionPolicy.RUNTIME)
+@Target({ElementType.METHOD, ElementType.TYPE})
+public @interface TestFlag {
+    boolean value() default false;
+}
+
